01_ecommerce_hourly_email_report/requirements.txt
pandas==2.2.2
pyarrow==17.0.0
yagmail==0.15.293
python-dotenv==1.0.1
